Struct serde_lexpr::parse::Options

pub struct Options { /* fields omitted */ }

Various options to influence parser behavior.

Methods

impl Options

pub fn new() -> Options

Construct an empty set of options.

This will recognize no keywords, and parse nil and t as regular symbols.

pub fn elisp() -> Options

Construct a set of options suitable for parsing Emacs Lisp.

pub fn with_keyword_syntax(self, syntax: KeywordSyntax) -> Options

Add syntax to the recognized keyword syntaxes.

pub fn with_keyword_syntaxes<I, T>(self, styles: I) -> Options where
I: IntoIterator<Item = T>,
T: Borrow<KeywordSyntax>,

Set the recognized keyword syntaxes.

pub fn with_nil_symbol(self, treatment: NilSymbol) -> Options

Choose how to parse the nil symbol.

pub fn with_t_symbol(self, treatment: TSymbol) -> Options

Choose how to parse the t symbol.

pub fn with_brackets(self, treatment: Brackets) -> Options

Choose how to handle brackets.

pub fn with_string_syntax(self, syntax: StringSyntax) -> Options

Choose the accepted string syntax.

pub fn with_char_syntax(self, syntax: CharSyntax) -> Options

Choose the accepted character syntax.

pub fn keyword_syntax(self, syntax: KeywordSyntax) -> bool

Check wether a keyword syntax is enabled.

pub fn nil_symbol(self) -> NilSymbol

Query the way the nil symbol is handled.

pub fn t_symbol(self) -> TSymbol

Query the way the t symbol is handled.

pub fn brackets(self) -> Brackets

Query the way brackets are handled.

pub fn string_syntax(self) -> StringSyntax

Query the accepted string syntax.

pub fn char_syntax(self) -> CharSyntax

Query the accepted character syntax.