Struct serde_lexpr::Cons

pub struct Cons { /* fields omitted */ }

A Lisp "cons cell".

A cons cell is similiar to a two-element tuple in Rust. Its fields are traditionally called car and cdr, for obscure historical reasons. Both the car and the cdr field can hold any Value, including other cons cells.

This data type is used to represent singly-linked lists, by forming a chain of cons cells where the list element is kept in the car field, and the cdr field either points to the next cons cell, or terminates the list with any other value. Usually, that terminator value is Value::Null, also referred to as the empty list. If any other terminating value is used, the resulting linked list is referred to as "dotted", or "improper" list.

The Cons data type provides some utility function for the singly-linked list use case, such as iterating through the list or converting the list to a vector. To account for the possibility of dotted lists, the iterators and vector conversion functions have slightly unusual types.

Methods

impl Cons

pub fn new<T, U>(car: T, cdr: U) -> Cons where
T: Into<Value>,
U: Into<Value>,

Constructs a new cons cell from two values.

pub fn car(&self) -> &Value

Returns a reference to the value in the car field.

pub fn car_mut(&mut self) -> &mut Value

Returns a mutable reference to the value in the car field.

pub fn set_car(&mut self, car: impl Into<Value>)

Sets the car field.

pub fn cdr(&self) -> &Value

Returns a reference to the value in the cdr field.

pub fn cdr_mut(&mut self) -> &mut Value

Returns a mutable reference to the value in the cdr field.

pub fn set_cdr(&mut self, cdr: impl Into<Value>)

Sets the cdr field.

pub fn as_pair(&self) -> (&Value, &Value)

Returns references to the values in the car and cdr fields.

let cell = Cons::new(1, 2);
assert_eq!(cell.as_pair(), (&Value::from(1), &Value::from(2)));

pub fn into_pair(self) -> (Value, Value)

Converts self into a pair of values without cloning.

let cell = Cons::new("a", 42);
assert_eq!(cell.car(), "a");
assert_eq!(cell.cdr(), 42);
let (car, cdr) = cell.into_pair();
assert_eq!(car, "a");
assert_eq!(cdr, 42);

pub fn iter(&self) -> Iter

Obtains an iterator yielding references to all the cons cells in this linked list.

for cell in Cons::new(1, Cons::new(2, Value::Null)).iter() {
   println!("list element: {}", cell.car());
}

pub fn into_vec(self) -> (Vec<Value>, Value)

Converts self into a vector without cloning the elements.

Returns the accumulated items of the list and the cdr of the last list element. For proper lists, this will always be Value::Null.

let list = Cons::new(1, Cons::new(2, Cons::new(3, Value::Null)));
assert_eq!(list.into_vec(), (vec![Value::from(1), Value::from(2), Value::from(3)], Value::Null));

pub fn to_vec(&self) -> (Vec<Value>, Value)

Retrieves a vector, cloning the values.

Returns the accumulated items of the list and the cdr of the last list element. For proper lists, this will always be Value::Null.

let list = Cons::new(1, Cons::new(2, Cons::new(3, Value::Null)));
assert_eq!(list.to_vec(), (vec![Value::from(1), Value::from(2), Value::from(3)], Value::Null));

pub fn to_ref_vec(&self) -> (Vec<&Value>, &Value)

Retrieves a vector, taking references to the values.

Returns the accumulated items of the list and the cdr of the last list element. For proper lists, this will always be Value::Null.

let list = Cons::new(1, Cons::new(2, Cons::new(3, Value::Null)));
assert_eq!(list.to_ref_vec(), (vec![&Value::from(1), &Value::from(2), &Value::from(3)], &Value::Null));