Trait serde_lexpr::parse::Read

pub trait Read<'de>: Sealed {
    fn parse_elisp_char(&mut self, scratch: &mut Vec<u8>) -> Result<char, Error> { ... }
}

Trait used by the parser for iterating over input.

This trait is manually "specialized" for iterating over &[u8]. Once feature(specialization) is stable we can use actual specialization.

This trait is sealed and cannot be implemented for types outside of lexpr.

Provided methods

fn parse_elisp_char(&mut self, scratch: &mut Vec<u8>) -> Result<char, Error>

Parses an Emacs Lisp character constant.

Loading content...